+++ src/styles/components/forms.css
@@ -4,13 +4,12 @@
   /* ========== Input/Select (4-state) ========== */
   .ui-input,
   .ui-select {
-    @apply w-full h-12 rounded-lg bg-white border px-4 text-gray-900 transition-colors duration-200;
-    border-color: #D1D5DB; /* As per Figma */
+    @apply w-full h-12 rounded-lg bg-white border px-4 text-gray-900 transition-colors duration-200 border-gray-300;
   }
   
   .ui-input::placeholder,
   .ui-select::placeholder {
-    color: #949AA6;
+        @apply text-gray-400;
   }
   
   .ui-input:hover,
