+++ templates/components/icon.html
@@ -13,73 +13,70 @@
 
 {% if name == "eye" %}
   <!-- 眼睛圖示 - 密碼可見 -->
-  <svg class="icon {{ size|default:'w-6 h-6' }} {{ class|default:'' }}" 
-       fill="none" 
-       stroke="currentColor" 
-       viewBox="0 0 24 24"
-       stroke-width="1.5"
-       stroke-linecap="round" 
-       stroke-linejoin="round">
-    <path d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"/>
-    <path d="M2.458 12C3.732 7.943 7.523 5 12 5c4.478 0 8.268 2.943 9.542 7-1.274 4.057-5.064 7-9.542 7-4.477 0-8.268-2.943-9.542-7z"/>
+  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M2.03613 12.322C1.96712 12.1146 1.96712 11.8904 2.03613 11.683C3.42313 7.51 7.36013 4.5 12.0001 4.5C16.6381 4.5 20.5731 7.507 21.9631 11.678C22.0331 11.885 22.0331 12.109 21.9631 12.317C20.5771 16.49 16.6401 19.5 12.0001 19.5C7.36213 19.5 3.42613 16.493 2.03613 12.322Z" stroke="#5D6970" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
+      <path d="M15 12C15 12.7956 14.6839 13.5587 14.1213 14.1213C13.5587 14.6839 12.7956 15 12 15C11.2044 15 10.4413 14.6839 9.87868 14.1213C9.31607 13.5587 9 12.7956 9 12C9 11.2044 9.31607 10.4413 9.87868 9.87868C10.4413 9.31607 11.2044 9 12 9C12.7956 9 13.5587 9.31607 14.1213 9.87868C14.6839 10.4413 15 11.2044 15 12Z" stroke="#5D6970" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
   </svg>
+
 
 {% elif name == "eye-off" %}
   <!-- 眼睛圖示 - 密碼隱藏 -->
-  <svg class="icon {{ size|default:'w-6 h-6' }} {{ class|default:'' }}" 
-       fill="none" 
-       stroke="currentColor" 
-       viewBox="0 0 24 24"
-       stroke-width="1.5"
-       stroke-linecap="round" 
-       stroke-linejoin="round">
-    <path d="M13.875 18.825A10.05 10.05 0 0112 19c-4.478 0-8.268-2.943-9.543-7a9.97 9.97 0 011.563-3.029m5.858.908a3 3 0 114.243 4.243M9.878 9.878l4.242 4.242M9.878 9.878L3 3m6.878 6.878L21 21"/>
+  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <g clip-path="url(#clip0_714_448)">
+      <path d="M3.97959 8.223C3.04405 9.32718 2.34741 10.6132 1.93359 12C3.22559 16.338 7.24359 19.5 11.9996 19.5C12.9926 19.5 13.9526 19.362 14.8626 19.105M6.22759 6.228C7.94024 5.09786 9.94768 4.49688 11.9996 4.5C16.7556 4.5 20.7726 7.662 22.0646 11.998C21.3566 14.3673 19.8366 16.4116 17.7716 17.772M6.22759 6.228L2.99959 3M6.22759 6.228L9.87759 9.878M17.7716 17.772L20.9996 21M17.7716 17.772L14.1216 14.122C14.4002 13.8434 14.6212 13.5127 14.772 13.1486C14.9227 12.7846 15.0003 12.3945 15.0003 12.0005C15.0003 11.6065 14.9227 11.2164 14.772 10.8524C14.6212 10.4883 14.4002 10.1576 14.1216 9.879C13.843 9.6004 13.5122 9.3794 13.1482 9.22863C12.7842 9.07785 12.3941 9.00025 12.0001 9.00025C11.6061 9.00025 11.216 9.07785 10.8519 9.22863C10.4879 9.3794 10.1572 9.6004 9.87859 9.879M14.1206 14.121L9.87959 9.88" stroke="#5D6970" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
+      </g>
+      <defs>
+      <clipPath id="clip0_714_448">
+      <rect width="24" height="24" fill="white"/>
+      </clipPath>
+      </defs>
   </svg>
+
 
 {% elif name == "google" %}
   <!-- Google圖示 - Google登入/註冊 -->
   <svg class="icon {{ size|default:'w-6 h-6' }} {{ class|default:'' }}" 
-       viewBox="0 0 24 24">
-    <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-    <path fill="#34A853" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-    <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
-    <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
+      viewBox="0 0 24 24">
+      <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
+      <path fill="#34A853" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
+      <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
+      <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
   </svg>
 
 {% elif name == "smart-recommend" %}
   <!-- 智慧推薦系統 -->
   <svg class="icon {{ size|default:'w-6 h-6' }} {{ class|default:'' }}" 
-       fill="none" 
-       stroke="currentColor" 
-       viewBox="0 0 24 24"
-       stroke-width="2"
-       stroke-linecap="round" 
-       stroke-linejoin="round">
-    <path d="M9.663 17h4.673M12 3v1m6.364 1.636l-.707.707M21 12h-1M4 12H3m3.343-5.657l-.707-.707m2.828 9.9a5 5 0 117.072 0l-.548.547A3.374 3.374 0 0014 18.469V19a2 2 0 11-4 0v-.531c0-.895-.356-1.754-.988-2.386l-.548-.547z"/>
+      fill="none" 
+      stroke="currentColor" 
+      viewBox="0 0 24 24"
+      stroke-width="2"
+      stroke-linecap="round" 
+      stroke-linejoin="round">
+      <path d="M9.663 17h4.673M12 3v1m6.364 1.636l-.707.707M21 12h-1M4 12H3m3.343-5.657l-.707-.707m2.828 9.9a5 5 0 117.072 0l-.548.547A3.374 3.374 0 0014 18.469V19a2 2 0 11-4 0v-.531c0-.895-.356-1.754-.988-2.386l-.548-.547z"/>
   </svg>
 
 {% elif name == "real-time-offer" %}
   <!-- 即時優惠提醒 -->
   <svg class="icon {{ size|default:'w-6 h-6' }} {{ class|default:'' }}" 
-       fill="none" 
-       stroke="currentColor" 
-       viewBox="0 0 24 24"
-       stroke-width="2"
-       stroke-linecap="round" 
-       stroke-linejoin="round">
-    <path d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"/>
+      fill="none" 
+      stroke="currentColor" 
+      viewBox="0 0 24 24"
+      stroke-width="2"
+      stroke-linecap="round" 
+      stroke-linejoin="round">
+      <path d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"/>
   </svg>
 
 {% elif name == "consumption-analysis" %}
   <!-- 消費分析報告 -->
   <svg class="icon {{ size|default:'w-6 h-6' }} {{ class|default:'' }}" 
-       fill="none" 
-       stroke="currentColor" 
-       viewBox="0 0 24 24"
-       stroke-width="2"
-       stroke-linecap="round" 
-       stroke-linejoin="round">
-    <path d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"/>
+      fill="none" 
+      stroke="currentColor" 
+      viewBox="0 0 24 24"
+      stroke-width="2"
+      stroke-linecap="round" 
+      stroke-linejoin="round">
+      <path d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"/>
   </svg>
 
 {% else %}
